python/pyproject.toml
[tool.poetry]
name = "tejai"
version = "0.0.1.dev"
description = "Tejai Python SDK"
authors = ["OLTEJ <product@oltej.com>"]
readme = "pip/README.md"
packages = [{include = "tejai"}]

[tool.poetry.dependencies]
python = "^3.8"


[tool.poetry.group.dev.dependencies]
pre-commit = "3.5.0"
# please keep black and ruff version in sync with the ones in .pre-commit-config.yaml
black = "^23.12.0"
ruff = "0.1.8"
ipykernel = "^6.21.1"
pytest = "7.4.3"
pytest-asyncio = "0.23.2"
snoop = "0.4.3"
pytest-cov = "4.1.0"
pyyaml = "6.0.1"

[tool.poetry.group.google_palm.dependencies]
google-generativeai = { version = ">=0.1,<0.3", markers = "python_version >= '3.9'" }
grpcio-status = { version = "^1.53.0", markers = "python_version >= '3.9'" }

[tool.poetry.group.hugging_face.dependencies]
transformers = "^4.28.1"
sentence-transformers = "^2.2.2"
torch = "2.0.0"

[tool.poetry.group.qdrant.dependencies]
qdrant-client = {version = "^1.3.2", python = ">=3.8,<3.12"}

[tool.poetry.group.chromadb.dependencies]
chromadb = "0.4.13"

[tool.poetry.group.milvus.dependencies]
pymilvus = "2.2.16"
milvus = "2.2.16"

[tool.poetry.group.weaviate.dependencies]
weaviate-client = "^3.18.0"

[tool.poetry.group.pinecone.dependencies]
pinecone-client = "^2.2.2"

[tool.poetry.group.postgres.dependencies]
psycopg-pool = "^3.1.7"
psycopg = "^3.1.9"
psycopg-binary = "^3.1.9"

[tool.poetry.group.redis.dependencies]
redis = "^4.6.0"

[tool.poetry.group.azure_cognitive_search.dependencies]
azure-search-documents = {version = "11.4.0b9", allow-prereleases = true}
azure-core = "^1.28.0"
azure-identity = "^1.13.0"

[tool.poetry.group.usearch.dependencies]
usearch = "1.1.1"
pyarrow = ">=12.0.1,<15.0.0"

[tool.isort]
profile = "ruff"

[tool.ruff]
select = ["E", "F", "I"]
line-length = 120

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"